feat : 소셜로그인 info 수정 #8
